backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==1.4.53
databases
psycopg2-binary==2.9.6
pydantic==1.10.7
python-multipart==0.0.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pdfplumber==0.7.5
python-docx==0.8.11
tqdm==4.65.0
python-dotenv==1.0.0
aiofiles==23.1.0
cached-property==1.5.2
requests==2.31.0
